Feature/db

+++ src/main/java/com/amalitech/demo/dto/request/CartItemRequest.java
@@ -1,7 +1,7 @@
 package com.amalitech.demo.dto.request;
 
-public record CartItemRequest(int cartId,
-    int productId,
+public record CartItemRequest(Long cartId,
+    Long productId,
     int quantity,
     Double unitPrice, Double totalPrice){
 }
